RaspberryPiV2/V2/backend/requirements.txt
flask
flask-cors
pyserial
psycopg2-binary
python-dotenv
pyopenssl
gunicorn
requests